requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
streamlit==1.29.0
Pillow==10.1.0
requests==2.31.0
